rename and update hooks accordingly, and add a new custom hook
