+++ client/src/hooks/use-filter-date.ts
@@ -1,4 +1,4 @@
-import { useFilterStore } from '@/lib/stores/filter.store';
+import { useFilterStore } from '@/stores/filter.store';
 
 export const useFilterDate = (field: 'startDate' | 'endDate') => {
   const filters = useFilterStore((state) => state.filters);
